Adding matrix multiplication and a Benchmark Test (#47)
* Not Functional Yet

* Added Matrix Multiplication

Changed all matrices to long inorder to account for larger numbers due to matrix multiplication

* Fixed Multiplication Implementation, and Added Benchmark Test

* Checkstyle Fixes

* Fix Checkstyle

* Fixed Slow/Faster Matrix Multiplication implementation

Got rid of extra classes and converted them into methods.

* Fixed Checkstyle

+++ src/src/DataStorage.java
@@ -7,7 +7,7 @@ public interface DataStorage{
   
   public void writeOutput(String outputFileName, String delimiter);
 
-  List<int[][]> readInputFile() throws IOException;
+  List<long[][]> readInputFile() throws IOException;
 
 
 }
