+++ test/tests/TestMatrix.java
@@ -21,10 +21,10 @@ public class TestMatrix {
   public void testMatrixImplementation() {
     MatrixImplementation matrixImplementation = new MatrixImplementation();
 
-    int[][] matrix1 = {{1, 2}, {3, 4}};
-    int[][] matrix2 = {{5, 6}, {7, 8}};
+    long[][] matrix1 = {{1, 2}, {3, 4}};
+    long[][] matrix2 = {{5, 6}, {7, 8}};
 
-    int[][] resultantMatrix = matrixImplementation.multiplyMatrices(matrix1, matrix2);
+    long[][] resultantMatrix = matrixImplementation.multiplyMatrices(matrix1, matrix2);
 
     assertNotNull(resultantMatrix);
     assertEquals(19, resultantMatrix[0][0]);
